Chrome OS
The chrome operating system is a Linux-based OS, which is specifically works with web applications such such, chrome, google docs, and Gmail.  This operating system is probably one of the cheaper versions of operating systems, mostly because you can only this this OS specifically for google products. Also, the chrome OS is only compatible for the internet. The Chrome OS only came out in 2011, so Its relatively new, but I think because this is such an affordable laptop its increasingly popular.